backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.0
google-genai==0.8.0
google-cloud-aiplatform==1.60.0
google-cloud-storage==2.18.0
google-cloud-texttospeech==2.16.0
firebase-admin==6.5.0
httpx==0.27.0
websockets==13.1
python-multipart==0.0.9